Adding functionality as well as closing capability to login/signUp modals
