+++ src/reducers/Actions.js
@@ -1,19 +1,16 @@
-module.exports = () => {
-    let i = 0;
+let i = 0;
 
-    return {
-        login: {
-            setEmail: i++,
-            setPassword: i++,
-            setFirstName: i++,
-            setLastName: i++,
-            setPasswordRepeat: i++,
-            setAddress: i++,
-            setCity: i++,
-            setAddressState: i++,
-            setZipcode: i++,
-            setDriversLiscenseNumber: i++,
-            clearPassword: i++
-        }
-    };
+export default {
+    login: {
+        setEmail: i++,
+        setPassword: i++,
+        setFirstName: i++,
+        setLastName: i++,
+        setPasswordRepeat: i++,
+        setAddress: i++,
+        setCity: i++,
+        setAddressState: i++,
+        setZipcode: i++,
+        clearPassword: i++
+    }
 };
